// pharmacy-ui/src/app/services/models/prescription-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PrescriptionItem } from '../models/prescription-item';
import { Users } from '../models/users';
export interface PrescriptionResponse {
  id?: number;
  image?: string;
  prescriptionItem?: Array<PrescriptionItem>;
  status?: 'PENDING' | 'APPROVED' | 'REJECTED';
  uploadedAt?: string;
  user?: Users;
}