// pharmacy-ui/src/app/services/models/prescription-item.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Prescriptions } from '../models/prescriptions';
import { Product } from '../models/product';
export interface PrescriptionItem {
  id?: number;
  prescriptions?: Prescriptions;
  product?: Product;
  quantity?: number;
}